prediction/models/NERMicroservice/requirements.txt
pillow
requests
rq
redis
tensorflow
transformers